Flatten CSS selector hierarchy a bit to improve style calculation times, especially for large files
While not completely flat, the rules are now:

  - less hierarchical
  - className- vs attribute-based selectors

This reduces the style calculation times to a more reasonable duration. If style calc performance is
still insufficient, I'll invest more time in completely flattening these rules, and try to opt for
more compositor-only CSS where possible, but it's a bigger change on the JS side.

[gocd_sf_team/#29] Optimize foldable section UI performance for large console logs (20,000+ lines)