Alternative implementation that speeds up expand/collapse section, but sacrifices first page load.
Instead of setting display:none to log lines in collapse sections, use transform: scale(0) in conjunction with height:0 and overflow-y: hidden.
As scale() is a compositor only property, changing this value in response to user input will not cause re-layout/reflow on each line, which is
the major bottleneck in expand/contract performance. The effect is significant (about an order of magnitude in a 20,000 line file).

However, as we are no longer using display:none, all log lines are part of the render tree, and thus we sacrifice significant performance
on page load. The difference in load time is more than noticeable (at the edge of tolerance), so this may not be a worthy tradeoff.